backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
mysqlclient==2.2.4
gunicorn==22.0.0
uvicorn==0.30.0
mercadopago==2.3.0
python-dateutil==2.9.0
 
channels==4.0.0
channels-redis==4.1.0
daphne==4.1.2
